fix warning e20 compiler finds with signed/unsigned compare.  Checked with LArPandoraTrackCreation_module.cc to make sure that the max int put in an unsigned int is truly intended.

+++ dune/Iceberg/Signal2Noise/Signal2Noise_module.cc
@@ -377,7 +377,10 @@ void Signal2Noise::analyze(art::Event const& e)
         for (size_t ii=0; ii<vhit.size(); ii++) {
           if (vhit[ii].key() == allhits[ihit].key()) {
             
-            if (vmeta[ii]->Index() == std::numeric_limits<int>::max()) {
+	    // nb.  LArPandoraTrackCreation_module.cc fills the max of a signed int in an unsigned int
+	    // to indicate an invalid index
+
+            if (vmeta[ii]->Index() == (unsigned int) std::numeric_limits<int>::max()) {
               fBadhit = true;
               continue;
             }
